Move message component types into the `model::application::component` module
The model types for message components are moved from
`interactions::message_component` into the `application::component` module.

Deprecated type aliases are kept in the original module.
