+++ examples/e17_message_components/src/main.rs
@@ -8,8 +8,8 @@ use serenity::async_trait;
 use serenity::builder::{CreateActionRow, CreateButton, CreateSelectMenu, CreateSelectMenuOption};
 use serenity::client::{Context, EventHandler};
 use serenity::futures::StreamExt;
+use serenity::model::application::component::ButtonStyle;
 use serenity::model::channel::Message;
-use serenity::model::interactions::message_component::ButtonStyle;
 use serenity::model::interactions::InteractionResponseType;
 use serenity::prelude::*;
